+++ test/eca/features/prompt_test.clj
@@ -8,7 +8,8 @@
   (testing "Should create instructions with rules, contexts, and behavior"
     (let [refined-contexts [{:type :file :path "foo.clj" :content "(ns foo)"}
                             {:type :file :path "bar.clj" :content "(def a 1)" :partial true}
-                            {:type :repoMap :path nil :content nil}]
+                            {:type :repoMap}
+                            {:type :mcpResource :uri "custom://my-resource" :content "some-cool-content"}]
           rules [{:name "rule1" :content "First rule"}
                  {:name "rule2" :content "Second rule"}]
           fake-repo-map (delay "TREE")
@@ -22,5 +23,6 @@
       (is (string/includes? result "<file path=\"foo.clj\">(ns foo)</file>"))
       (is (string/includes? result "<file partial=true path=\"bar.clj\">...\n(def a 1)\n...</file>"))
       (is (string/includes? result "<repoMap description=\"Workspaces structure in a tree view, spaces represent file hierarchy\" >TREE</repoMap>"))
+      (is (string/includes? result "<resource uri=\"custom://my-resource\">some-cool-content</resource>"))
       (is (string/includes? result "</contexts>"))
       (is (string? result)))))